move IDropdownOption interface in other folder and fix path issue
