+++ console/console-init/ui/src/components/common/index.ts
@@ -10,7 +10,6 @@ export * from "./DialogPrompt";
 export * from "./EmptyLinks";
 export * from "./Error";
 export * from "./ErrorBoundary";
-export * from "./FilterDropdown";
 export * from "./NotFound";
 export * from "./ServerMessageAlert";
 export * from "./SortForMobileView";
